# apps/web/openapi/generated/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "CourseLit REST API",
        "version": "1.0.0",
        "description": "OpenAPI documentation for the public CourseLit REST API surface."
    },
    "servers": [
        {
            "url": "/",
            "description": "Current CourseLit school origin"
        }
    ],
...
    "paths": {
...
        "/api/products/{productId}/lessons/{lessonId}": {
            "get": {
                "tags": ["Product Content"],
                "summary": "Get a product lesson",
                "operationId": "getProductLesson",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lessonId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lesson not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": ["Product Content"],
                "summary": "Update a product lesson",
                "description": "Updates editable lesson fields. Lesson type and section cannot be changed after creation. Use the same content/media shapes documented on create. SCORM lesson updates are rejected with `not_supported`.",
                "operationId": "updateProductLesson",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lessonId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LessonUpdateRequest"
                            },
                            "example": {
                                "title": "Updated Introduction to AI",
                                "published": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "SCORM lessons are not supported, or lesson validation failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Product Content"],
                "summary": "Delete a product lesson",
                "operationId": "deleteProductLesson",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lessonId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lesson could not be deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products/{productId}/customers": {
            "get": {
                "tags": ["Product Customers"],
                "summary": "List product customers",
                "operationId": "listProductCustomers",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search customers by name or email.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "minimum": 1,
                            "maximum": 200
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerListResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{productId}/customers/invitations": {
            "post": {
                "tags": ["Product Customers"],
                "summary": "Invite a customer",
                "description": "Invites a customer by email. An invitation email is sent to the provided address.",
                "operationId": "inviteProductCustomer",
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "required": ["email"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unsupported customer invitation field or missing email.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Customer could not be invited.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PublicApiErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "object",
                        "properties": {
                            "code": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "required": ["code", "message"]
                    }
                },
                "required": ["error"]
            },
            "OkResponse": {
                "type": "object",
                "properties": {
                    "ok": {
                        "type": "boolean"
                    }
                }
            },
...
            "TiptapDocument": {
                "type": "object",
                "description": "Tiptap/ProseMirror document JSON used by `text` lessons. Supported nodes include `doc`, `paragraph`, `heading` (`level` 1, 2, or 3), `text`, `bulletList`, `orderedList`, `listItem`, `blockquote`, `horizontalRule`, `codeBlock`, `table`, `tableRow`, `tableHeader`, `tableCell`, `image`, and `hardBreak`. Supported marks include `bold`, `italic`, `underline`, `strike`, `code`, `link`, and `highlight`. An `image` node uses `attrs` with `src` (required), `alt` (optional), and `title` (optional).",
                "required": ["type", "content"],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "doc"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                }
            },
            "EmbedContent": {
                "type": "object",
                "description": "Embed content for `embed` lessons. The value can be a supported video URL or iframe/embed code, matching the dashboard Embed lesson field.",
                "required": ["value"],
                "properties": {
                    "value": {
                        "type": "string",
                        "example": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                    }
                }
            },
            "QuizContent": {
                "type": "object",
                "description": "Quiz content for `quiz` lesson type.",
                "required": [
                    "questions",
                    "requiresPassingGrade",
                    "passingGrade"
                ],
                "properties": {
                    "questions": {
                        "type": "array",
                        "description": "List of quiz questions.",
                        "items": {
                            "type": "object",
                            "required": ["text", "options"],
                            "properties": {
                                "text": {
                                    "type": "string",
                                    "description": "Question text."
                                },
                                "options": {
                                    "type": "array",
                                    "description": "Answer options. Exactly one option should have `correctAnswer: true` for single-choice; multiple for multiple-choice.",
                                    "items": {
                                        "type": "object",
                                        "required": ["text"],
                                        "properties": {
                                            "text": {
                                                "type": "string",
                                                "description": "Option text."
                                            },
                                            "correctAnswer": {
                                                "type": "boolean",
                                                "description": "Whether this is the correct answer. Stripped from student-facing responses."
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "requiresPassingGrade": {
                        "type": "boolean",
                        "description": "Whether a minimum score is required to pass."
                    },
                    "passingGrade": {
                        "type": "number",
                        "description": "Score threshold (0–100) when `requiresPassingGrade` is true."
                    }
                }
            },
            "LessonMedia": {
                "type": "object",
                "description": "Media object used by `video`, `audio`, `pdf`, and `file` lessons. Use `/api/media/presigned` to upload to MediaLit first, then send the resulting media metadata here.",
                "required": ["mediaId"],
                "properties": {
                    "mediaId": {
                        "type": "string"
                    },
                    "originalFileName": {
                        "type": "string"
                    },
                    "mimeType": {
                        "type": "string"
                    },
                    "size": {
                        "type": "number"
                    },
                    "access": {
                        "type": "string",
                        "enum": ["public", "private"]
                    },
                    "file": {
                        "type": "string",
                        "description": "Public file URL when available. Private media may omit this field."
                    },
                    "thumbnail": {
                        "type": "string"
                    },
                    "caption": {
                        "type": "string"
                    }
                }
            },
            "Lesson": {
                "type": "object",
                "properties": {
                    "lessonId": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "text",
                            "video",
                            "audio",
                            "pdf",
                            "file",
                            "embed",
                            "quiz"
                        ]
                    },
                    "content": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/TiptapDocument"
                            },
                            {
                                "$ref": "#/components/schemas/EmbedContent"
                            },
                            {
                                "$ref": "#/components/schemas/QuizContent"
                            }
                        ],
                        "description": "`text` lessons use `TiptapDocument`; `embed` lessons use `EmbedContent`; `quiz` lessons use `QuizContent`. Media-backed lessons (`video`, `audio`, `pdf`, `file`) use `media` instead of `content`."
                    },
                    "media": {
                        "$ref": "#/components/schemas/LessonMedia"
                    },
                    "downloadable": {
                        "type": "boolean"
                    },
                    "courseId": {
                        "type": "string"
                    },
                    "groupId": {
                        "type": "string"
                    },
                    "requiresEnrollment": {
                        "type": "boolean"
                    },
                    "published": {
                        "type": "boolean"
                    }
                }
            },
...
            "LessonUpdateRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/TiptapDocument"
                            },
                            {
                                "$ref": "#/components/schemas/EmbedContent"
                            },
                            {
                                "$ref": "#/components/schemas/QuizContent"
                            }
                        ],
                        "description": "`text` lessons use `TiptapDocument`; `embed` lessons use `EmbedContent`; `quiz` lessons use `QuizContent`. Media-backed lessons (`video`, `audio`, `pdf`, `file`) use `media` instead of `content`."
                    },
                    "media": {
                        "$ref": "#/components/schemas/LessonMedia"
                    },
                    "downloadable": {
                        "type": "boolean"
                    },
                    "requiresEnrollment": {
                        "type": "boolean"
                    },
                    "published": {
                        "type": "boolean"
                    }
                }
            },
...
            "Customer": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "object"
                    },
                    "membershipId": {
                        "type": "string"
                    },
                    "membershipStatus": {
                        "type": "string"
                    },
                    "subscriptionMethod": {
                        "type": "string"
                    },
                    "completedLessons": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "downloaded": {
                        "type": "boolean"
                    },
                    "enrolledAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ProductCustomer": {
                "type": "object",
                "description": "Member enrollment details returned by the product customers endpoint.",
                "properties": {
                    "user": {
                        "type": "object",
                        "properties": {
                            "userId": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "avatar": {
                                "type": "object"
                            }
                        }
                    },
                    "status": {
                        "type": "string"
                    },
                    "completedLessons": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "downloaded": {
                        "type": "boolean"
                    },
                    "subscriptionMethod": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CustomerListResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductCustomer"
                        }
                    },
                    "pagination": {
                        "type": "object",
                        "properties": {
                            "page": {
                                "type": "integer"
                            },
                            "limit": {
                                "type": "integer"
                            }
                        }
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key",
                "description": "API key created in CourseLit dashboard settings. The legacy `apikey` request-body field is still accepted but deprecated."
            }
        }
    }
}